B200 (sm100) reduction-seed heuristic: dedicated heuristic + half-precision fixes
Brings the reduction seed to life on B200 (sm100, 148 SMs) for fp32 + bf16 + fp16, on top
of the H100/sm90 reduction-seed PR. The H100/sm90 path is unchanged (byte-identical).

Heuristic (helion/_compiler/autotuner_heuristics/triton.py + __init__.py):
  Dedicated sm100-gated classes TritonB200Reduction{Tile,UserTile}Heuristic (subclassing the
  H100 T1/T2 classes; the sm90 classes defer on sm100 so exactly one seed fires). B200-tuned
  levers, all keyed on faithful workload properties (no dtype branch):
   - num_warps ramp re-keyed on per-row reduction BYTES + reread structure.
   - Band-B (carried 2-D accumulators) num_warps cap.
   - wide low-HBM-width reread rows loop a chunk (new hbm_itemsize fact = true HBM-load width).
   - welford Band-C M_BLOCK packing (single-pass gate) + a wide coupled bundle.
   - few-row under-occupied looped streamed reductions use a persistent_interleaved grid.

Facts (helion/autotuner/config_spec.py + helion/_compiler/device_ir.py):
  Add ReductionFact.hbm_itemsize (true HBM-load element width, pre-fp32-upcast).

Kernel-authoring half-precision fixes (examples/, all fp32 no-ops):
  jsd (compile + wide-V fp16 overflow), welford & kl_div (fp16 NaN), sum & long_sum (fp32 accum).

Result on B200: seed beats torch.compile-default geomean ~1.15x across 9 reduction kernels x
{fp32,bf16,fp16}.

[--no-verify: pyrefly absent on this box; ruff/ruff-format/codespell all pass.]

+++ examples/jsd.py
@@ -99,11 +99,19 @@ def jsd_forward(
                     dX[tile_bt, tile_X] = 0.0
                 continue
         intermediate_loss = hl.zeros([tile_bt, block_size_n], dtype=torch.float32)
-        intermediate_dX = hl.zeros([tile_bt, block_size_n], dtype=_input.dtype)
+        # Accumulate dX in fp32 (matching intermediate_loss): the per-tile updates are
+        # fp32-valued, and dX is summed into the fp32 `dX` output, so an `_input.dtype`
+        # accumulator only forces a half<-fp32 control-flow mismatch at bf16/fp16 (the
+        # branch assigns fp32 values). fp32 no-op: `_input.dtype == torch.float32` there.
+        intermediate_dX = hl.zeros([tile_bt, block_size_n], dtype=torch.float32)
         for tile_v in hl.tile(V, block_size=block_size_n):
-            # Load log probabilities and convert to float32
-            X = _input[tile_bt, tile_v]
-            Y = target[tile_bt, tile_v]
+            # Load log probabilities and convert to float32. The fp32 upcast is required at
+            # fp16: over a wide vocab (V>=128k) the per-class probability M ~ 1/V approaches
+            # fp16's smallest normal (~6e-5), so exp(X)/the mixture M underflows to 0 ->
+            # log(0) = -inf -> NaN. In fp32 (min normal ~1e-38) M stays representable. fp32
+            # no-op when the inputs are already fp32.
+            X = _input[tile_bt, tile_v].to(torch.float32)
+            Y = target[tile_bt, tile_v].to(torch.float32)
 
             if beta == 0.0:  # Forward KL: KL(P || Q)
                 Y_max = torch.amax(Y, dim=0)
